pyproject.toml
[tool.poetry]
name = "falcons-eye"
version = "0.1.0"
description = ""
authors = ["syafiq <ahmadsyafiq93@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
psycopg2-binary = "^2.8.6"
celery = {extras = ["redis"], version = "^5.1.0"}
requests = "^2.25.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
